intellidoc/requirements.txt
flask
dash
openai
PyMuPDF
pytesseract
Pillow
scikit-learn
numpy
ultralytics
opencv-python
faiss-cpu
python-dotenv
requests
gunicorn
pandas
